+++ pdf-generator/functions/__tests__/functions.test.ts
@@ -4,7 +4,7 @@ process.env.FIRESTORE_EMULATOR_HOST = "localhost:8080";
 process.env.FIREBASE_FIRESTORE_EMULATOR_ADDRESS = "localhost:8080";
 process.env.FIREBASE_AUTH_EMULATOR_HOST = "localhost:9099";
 
-admin.initializeApp({ projectId: "demo-project" });
+admin.initializeApp({ projectId: "demo-test" });
 
 describe("functions", () => {
   test("should generate a pdf", async () => {