+++ pdf-generator/functions/src/index.ts
@@ -1,36 +1,19 @@
-const { OUTPUT_STORAGE_BUCKET, RETURN_PDF_IN_RESPONSE, TEMPLATE_PATH } =
-  process.env as {
-    OUTPUT_STORAGE_BUCKET?: string;
-    RETURN_PDF_IN_RESPONSE: string;
-    TEMPLATE_PATH: string;
-  };
-
-const extensionParameters = {
-  OUTPUT_STORAGE_BUCKET,
-  RETURN_PDF_IN_RESPONSE,
-  TEMPLATE_PATH,
-};
-
+import { types as utilTypes } from "util";
 import * as functions from "firebase-functions";
 import * as admin from "firebase-admin";
-import { initializeApp } from "firebase/app";
-import {
-  connectStorageEmulator,
-  getDownloadURL,
-  getStorage,
-  ref,
-  uploadBytes,
-} from "firebase/storage";
-import { loadTemplate } from "./load_template";
-import { renderPdf } from "./render_pdf";
-import { serveTemplate } from "./serve_template";
 import { ParamsDictionary, Request } from "express-serve-static-core";
-import { GetParameters, parseParameters } from "./parse_parameters";
+import {
+  extensionParameters,
+  GetParameters,
+  ParsedParameters,
+  parseParameters,
+} from "./parse_parameters";
 import { getEventarc } from "firebase-admin/eventarc";
+import { producePdf } from "./produce_pdf";
 
 admin.initializeApp();
 
-const eventChannel = process.env.EVENTARC_CHANNEL
+export const eventChannel = process.env.EVENTARC_CHANNEL
   ? getEventarc().channel(process.env.EVENTARC_CHANNEL, {
       allowedEventTypes: process.env.EXT_SELECTED_EVENTS,
     })
@@ -40,181 +23,161 @@ process.on("unhandledRejection", (reason, p) => {
   console.error(reason, "Unhandled Rejection at Promise", p);
 });
 
-exports.executePdfGenerator = functions.handler.https.onRequest(
+let contextStack: string[];
+
+/**
+ * Wraps an action function and records its context for better error reporting
+ *
+ * @param {Function} action The action to run
+ * @return {Generic} The return value of the action
+ */
+export function runAction<Arguments, ReturnValue>(
+  action: (...args: Arguments[]) => ReturnValue,
+  ...args: Arguments[]
+): ReturnValue {
+  contextStack.push(
+    action.name
+      .split(/(?=[A-Z])/)
+      .map((word) => word.replace(/^./, (letter) => letter.toLowerCase()))
+      .join(" ")
+  );
+  const result = action(...args);
+  if (utilTypes.isPromise(result)) {
+    return new Promise((resolve, reject) => {
+      result.then(
+        (value) => {
+          contextStack.pop();
+          resolve(value);
+        },
+        (error) => {
+          reject(error);
+        }
+      );
+    }) as ReturnValue;
+  } else {
+    contextStack.pop();
+    return result;
+  }
+}
+/**
+ * Returns an error handler to be called whenever an exception is raised
+ *
+ * @return {Function} The error handler callback
+ */
+function createErrorHandler({
+  context,
+  response,
+}: {
+  context: { [key: string]: unknown };
+  response?: functions.Response<unknown>;
+}): (error: Error) => void {
+  contextStack = [];
+  return async function handleError(error: Error) {
+    functions.logger.error(error);
+    response?.status(500);
+    response?.setHeader("content-type", "application/json");
+    response?.end(
+      JSON.stringify({
+        error: error.message,
+        context: contextStack.join(" -> "),
+      })
+    );
+
+    if (eventChannel) {
+      await eventChannel.publish({
+        type: "firebase.extensions.pdf-generator.v1.error",
+        subject: "error",
+        data: { ...context, extensionParameters },
+      });
+    }
+  };
+}
+
+exports.executePdfGeneratorHttp = functions.handler.https.onRequest(
   async (
     request: Request<ParamsDictionary, unknown, unknown, GetParameters>,
     response
   ) => {
-    let context = "";
-
-    // eslint-disable-next-line require-jsdoc
-    function handleError(error: Error) {
-      console.error("Enexpected error");
-      console.error(error);
-      response.status(500);
-      response.setHeader("content-type", "application/json");
-      response.end(
-        JSON.stringify({
-          error: error.message,
-          context,
-        })
-      );
-    }
+    const errorHandler = createErrorHandler({
+      response,
+      context: {
+        query: request.query,
+      },
+    });
 
     try {
-      process.on("uncaughtException", handleError);
+      process.on("uncaughtException", errorHandler);
 
-      context = "parse-parameters";
-      functions.logger.info("Parsing get parameters");
-      const {
-        adjustHeightToFit,
-        chromiumPdfOptions,
-        data,
-        headful,
-        outputFileName,
-        templateBucket,
-        templatePrefix,
-        templateId,
-        shouldWaitForIsReady,
-      } = parseParameters({
-        query: request.query,
-        defaultTemplatePath: TEMPLATE_PATH,
-      });
-      functions.logger.info("Get parameters parsed successfully");
-      context = "";
-
-      functions.logger.info("Generating pdf from template", {
-        templateBucket,
-        templatePrefix,
-        templateId,
-      });
-
-      context = "initialize-firebase-storage";
-      functions.logger.info("Initializing Firebase Storage");
-      const firebaseConfig = {
-        storageBucket: templateBucket,
-      };
-      const app = initializeApp(firebaseConfig);
-      const storage = getStorage(app);
-      if (process.env.STORAGE_EMULATOR_PORT != null) {
-        connectStorageEmulator(
-          storage,
-          "127.0.0.1",
-          Number.parseInt(process.env.STORAGE_EMULATOR_PORT)
-        );
-      }
-      functions.logger.info("Firebase Storage initialized successfully");
-      context = "";
-
-      context = "load-template";
-      functions.logger.info("Loading template file");
-      const templateFilesPath = await loadTemplate({
-        storage,
-        templatePrefix,
-        templateId,
-        data,
-      });
-      functions.logger.info("Template file loaded successfully");
-      context = "";
-
-      context = "serve-template";
-      functions.logger.info("Serving template directory");
-      const portNumber = await serveTemplate({ path: templateFilesPath });
-      functions.logger.info("Template directory served successfully");
-      context = "";
-
-      context = "generate-pdf";
-      functions.logger.info("Generating pdf file");
-      const pdf = await renderPdf({
-        adjustHeightToFit,
-        chromiumPdfOptions,
-        headless: process.env.IS_LOCAL != "true" || !headful,
-        portNumber,
-        shouldWaitForIsReady,
-      });
-      functions.logger.info("Pdf file generated successfully");
-      context = "";
-
-      let publicUrl: string | undefined;
-
-      if (OUTPUT_STORAGE_BUCKET != null && OUTPUT_STORAGE_BUCKET != "") {
-        context = "upload-pdf";
-        functions.logger.info("Uploading pdf file to Firebase Storage");
-        const firebaseConfig = {
-          storageBucket: OUTPUT_STORAGE_BUCKET,
-        };
-        const app = initializeApp(firebaseConfig);
-        const storage = getStorage(app);
-        if (process.env.STORAGE_EMULATOR_PORT != null) {
-          connectStorageEmulator(
-            storage,
-            "127.0.0.1",
-            Number.parseInt(process.env.STORAGE_EMULATOR_PORT)
-          );
-        }
-        const pdfRef = ref(storage, outputFileName);
-        await uploadBytes(pdfRef, pdf);
-        functions.logger.info(
-          "Pdf file uploaded to Firebase Storage successfully"
-        );
-        publicUrl = await getDownloadURL(pdfRef);
-        context = "";
-      }
-
-      const pdfInformation = {
-        timestamp: new Date().getTime(),
-        publicUrl,
-        fileSize: pdf.length,
-        parameters: {
-          adjustHeightToFit,
-          chromiumPdfOptions,
-          data,
-          fileName: outputFileName,
-          templateBucket,
-          templatePrefix,
-          templateId,
-        },
+      const parameters = runAction(parseParameters, {
+        rawParameters: request.query,
         extensionParameters,
-      };
+      });
 
-      if (RETURN_PDF_IN_RESPONSE.toLowerCase() === "yes") {
+      const { pdf, functionContext } = await runAction(producePdf, {
+        outputBucketName: extensionParameters.OUTPUT_STORAGE_BUCKET,
+        parameters,
+      });
+
+      if (extensionParameters.RETURN_PDF_IN_RESPONSE.toLowerCase() === "yes") {
         response.setHeader(
           "content-type",
-          `application/pdf; filename="${outputFileName}"`
+          `application/pdf; filename="${parameters.outputFileName}"`
         );
         response.setHeader(
           "content-disposition",
-          `inline; filename="${outputFileName}"`
+          `inline; filename="${parameters.outputFileName}"`
         );
         response.end(pdf);
       } else {
         response.setHeader("content-type", "application/json");
-        response.end(JSON.stringify({ done: "successful", pdfInformation }));
-      }
-      if (eventChannel) {
-        await eventChannel.publish({
-          type: "firebase.extensions.pdf-generator.v1.complete",
-          subject: templateId,
-          data: pdfInformation,
-        });
+        response.end(JSON.stringify({ done: "successful", functionContext }));
       }
     } catch (error) {
-      if (eventChannel) {
-        await eventChannel.publish({
-          type: "firebase.extensions.pdf-generator.v1.error",
-          subject: TEMPLATE_PATH,
-          data: {
-            query: request.query,
-            extensionParameters,
-          },
-        });
-      }
       functions.logger.error("Error", error);
       if (error instanceof Error) {
-        handleError(error);
+        errorHandler(error);
       }
     } finally {
-      process.removeListener("uncaughtException", handleError);
+      process.removeListener("uncaughtException", errorHandler);
     }
   }
 );
+
+exports.executePdfGeneratorFirestore =
+  functions.handler.firestore.document.onCreate(
+    async (
+      snapshot: functions.firestore.QueryDocumentSnapshot,
+      context: functions.EventContext
+    ) => {
+      const id = snapshot.id;
+      const data = snapshot.data();
+      const errorHandler = createErrorHandler({
+        context: {
+          id,
+          data,
+          context,
+        },
+      });
+
+      try {
+        process.on("uncaughtException", errorHandler);
+
+        const parameters: ParsedParameters = runAction(parseParameters, {
+          rawParameters: { data, id },
+          extensionParameters,
+        });
+
+        await runAction(producePdf, {
+          outputBucketName: extensionParameters.OUTPUT_STORAGE_BUCKET,
+          parameters,
+        });
+      } catch (error) {
+        functions.logger.error("Error", error);
+        if (error instanceof Error) {
+          errorHandler(error);
+        }
+      } finally {
+        process.removeListener("uncaughtException", errorHandler);
+      }
+    }
+  );