+++ pdf-generator/functions/src/serve_template.ts
@@ -1,15 +1,17 @@
 import http from "http";
+import * as functions from "firebase-functions";
 
 import serveStatic from "serve-static";
 
 /**
- * Serves a directory on a free port in localhost
+ * Serves a directory on a free port in localhost.
  */
 export async function serveTemplate({
   path,
 }: {
   path: string;
 }): Promise<number> {
+  functions.logger.info("Serving template directory");
   const serve = serveStatic(path, { dotfiles: "allow", fallthrough: false });
 
   const server = http.createServer(function (request, response) {
@@ -24,6 +26,8 @@ export async function serveTemplate({
   const portNumber = await getPort();
 
   server.listen(portNumber);
+
+  functions.logger.info("Template directory served successfully");
 
   return portNumber;
 }